Check library isochrones for NaNs
After `mklib` is run, each installed isochrone is checked for
occurences of `nan` values. These are consistent with some as-yet
unknown interpolation error in `mklib` with some of the newer Padova
isochrones. The course of action is to find these isochrones and
remove them from the isofile so that they do not trip up `synth`.