Synth plots Hess diagrams of synthetic CMDs
The `Synth.plot_all_hess` method is a wrapper for automating the
plotting of all available synthetic CMDs. The actually plotting
is done by `Synth._plot_hess`.

If synthetic CMD file is not available, and error is logged
and the `plot_all_hess` wrapper continues. I don't know why
some entires in the `isofile` are being skipped.