Refactor LibraryBuilder parameters
All parameters for lib.dat are now set in `LibraryBuilder.__init__`.
This allows us to create a valid `LibraryBuilder` instance without
having to run `mklib` itseful. This is useful because `LibraryBuilder`
is used by other classes as storage for paths and parameters.
Paths to the isofile and lib.dat are also set at `__init__`.