Create SFHCirclePlot
This plot class represents star formation rate amplitudes in the
age/Z plane as circular areas. This type of plot is inspired by
Brown's papers from the HST/ACS survey of M31. The plot can
either be inserted into a user-provided axes instance, or a full
plot can be generated and saved to disk.